feat-chore: openai_chat package to manage code
