+++ autobots/action/action_chat/chat_crud.py
@@ -7,7 +7,7 @@ from pymongo.database import Database
 from pymongo.results import DeleteResult
 
 from autobots.action.action_chat.chat_doc_model import ChatDoc, ChatDocCreate, ChatDocFind, ChatDocUpdate
-from autobots.conn.openai.chat import Message
+from autobots.conn.openai.openai_chat.chat_model import Message
 from autobots.core.database.mongo_base import get_mongo_db
