+++ autobots/action/action_type/action_text2text/action_text2text_llm_chat_openai_v2.py
@@ -1,12 +1,12 @@
-from typing import Optional, Dict, Any, Type
+from typing import Optional, Type
 
 from pydantic import ValidationError
 
 from autobots.action.action_type.abc.IAction import IAction, ActionOutputType, ActionInputType, ActionConfigType
-from autobots.action.action.action_doc_model import ActionCreate, ActionDoc
+from autobots.action.action.action_doc_model import ActionCreate
 from autobots.action.action_type.action_types import ActionType
 from autobots.action.action.common_action_models import TextObj, TextObjs
-from autobots.conn.openai.chat import Message, ChatReq, Role
+from autobots.conn.openai.openai_chat.chat_model import Message, ChatReq, Role
 from autobots.conn.openai.openai_client import get_openai
 from autobots.core.log import log
 
@@ -42,7 +42,7 @@ class ActionGenTextLlmChatOpenaiV2(IAction[ChatReq, TextObj, TextObjs]):
             if action_input and action_input.text != "":
                 message = Message(role=Role.user, content=action_input.text)
                 self.action_config.messages = self.action_config.messages + [message]
-            chat_res = await get_openai().chat(chat_req=self.action_config)
+            chat_res = await get_openai().openai_chat.chat(chat_req=self.action_config)
             if not chat_res:
                 return text_objs
             # resp = Message.model_validate(chat_res.choices[0].message)
