+++ autobots/conn/openai/openai_chat/chat_model.py
