+++ autobots/conn/openai/openai_client.py
@@ -1,14 +1,12 @@
-import time
 from functools import lru_cache
 
-from openai import AsyncOpenAI, AsyncStream
+from openai import AsyncOpenAI
 from openai.types import CreateEmbeddingResponse, ImagesResponse
-from openai.types.chat import ChatCompletion, ChatCompletionChunk
 
-from autobots.conn.openai.chat import ChatReq
 from autobots.conn.openai.embedding import EmbeddingReq, EmbeddingRes
 from autobots.conn.openai.image_model import ImageReq
 from autobots.conn.openai.openai_audio.openai_audio import OpenaiAudio
+from autobots.conn.openai.openai_chat.openai_chat import OpenaiChat
 from autobots.core.log import log
 from autobots.core.settings import Settings, SettingsProvider
 
@@ -16,35 +14,9 @@ from autobots.core.settings import Settings, SettingsProvider
 class OpenAI:
 
     def __init__(self, org_id: str, api_key: str):
-        self.client = AsyncOpenAI(
-            api_key=api_key,
-            organization=org_id,
-            timeout=180
-        )
+        self.client = AsyncOpenAI(api_key=api_key,organization=org_id,timeout=180)
         self.openai_audio = OpenaiAudio(self.client)
-
-
-    async def chat(self, chat_req: ChatReq) -> ChatCompletion | AsyncStream[ChatCompletionChunk] | None:
-        max_retry = 3
-        for i in range(max_retry):
-            try:
-                log.trace("Starting OpenAI Chat, try: 1")
-                res: ChatCompletion = await self.client.chat.completions.create(**chat_req.model_dump(), timeout=180)
-                log.trace("Completed OpenAI Chat")
-                if isinstance(res, AsyncStream):
-                    return self.yield_chat_chunks(res)
-                else:
-                    return res
-            except Exception as e:
-                log.exception(e)
-                time.sleep(5)
-
-    async def yield_chat_chunks(self, chat_res: AsyncStream[ChatCompletionChunk]) -> ChatCompletionChunk | None:
-        try:
-            async for part in chat_res:
-                yield part
-        except Exception as e:
-            log.exception(e)
+        self.openai_chat = OpenaiChat(self.client)
 
     async def embedding(self, embedding_req: EmbeddingReq) -> EmbeddingRes | None:
         try:
