+++ tests/conn/openai_client/test_chat.py
@@ -1,9 +1,7 @@
 import pytest as pytest
-from openai import AsyncStream
-from openai.types.chat import ChatCompletion, ChatCompletionUserMessageParam, ChatCompletionSystemMessageParam, \
-    ChatCompletionChunk
+from openai.types.chat import ChatCompletion, ChatCompletionUserMessageParam, ChatCompletionSystemMessageParam
 
-from autobots.conn.openai.chat import ChatReq
+from autobots.conn.openai.openai_chat.chat_model import ChatReq
 from autobots.conn.openai.openai_client import get_openai
 
 
@@ -13,7 +11,7 @@ async def test_chat_happy_path(set_test_settings):
     msg1 = ChatCompletionUserMessageParam(role="user", content="Most famous Mechanics law")
     params = ChatReq(messages=[msg0, msg1], model="gpt-3.5-turbo-16k-0613")
 
-    resp: ChatCompletion = await get_openai().chat(chat_req=params)
+    resp: ChatCompletion = await get_openai().openai_chat.chat(chat_req=params)
 
     assert "assistant" == resp.choices[0].message.role
     assert "Newton" in resp.choices[0].message.content
@@ -25,7 +23,7 @@ async def test_chat_stream_happy_path(set_test_settings):
     msg1 = ChatCompletionUserMessageParam(role="user", content="Most famous Mechanics law")
     params = ChatReq(messages=[msg0, msg1], model="gpt-3.5-turbo-16k-0613", stream=True)
 
-    stream = await get_openai().chat(chat_req=params)
+    stream = await get_openai().openai_chat.chat(chat_req=params)
 
     collected_words = ""
     async for part in stream:
